+++ Runtime/ColorSet.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace UnderLogic.Collections
+{
+    [CreateAssetMenu(menuName = "Collections/Sets/Color Set")]
+    public class ColorSet : RuntimeSet<Color> { }
+}