+++ Runtime/Vector2IntSet.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace UnderLogic.Collections
+{
+    [CreateAssetMenu(menuName = "Collections/Sets/Vector2Int Set")]
+    public class Vector2IntSet : RuntimeSet<Vector2Int> { }
+}