+++ CHANGELOG.md
@@ -8,6 +8,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `Add(int)` method to `FloatSet` collection
+- `Remove(int)` method to `FloatSet` collection
+- `Add(int)` method to `DoubleSet` collection
+- `Add(float)` method to `DoubleSet` collection
+- `Remove(int)` method to `DoubleSet` collection
+- `Remove(float)` method to `DoubleSet` collection
+- `Add(bool)` method to `StringSet` collection
+- `Add(int)` method to `StringSet` collection
+- `Add(float)` method to `StringSet` collection
+- `Add(double)` method to `StringSet` collection
+- `Remove(bool)` method to `StringSet` collection
+- `Remove(int)` method to `StringSet` collection
+- `Remove(float)` method to `StringSet` collection
+- `Remove(double)` method to `StringSet` collection
+- `Add(Vector2Int)` method to `Vector2Set` collection
+- `Remove(Vector2Int)` method to `Vector2Set` collection
+- `Vector2IntSet` collection
+- `ColorSet` collection
+
+## [0.1.1] - 2022-12-01
+
+### Added
+
 - `IntSet` collection
 - `FloatSet` collection
 - `DoubleSet` collection